pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "terrametria"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "renardeinside", email = "polarpersonal@gmail.com" }]
dependencies = ["geopandas"]

[project.optional-dependencies]
server = ["fastapi", "uvicorn", "databricks-sql-connector", "python-dotenv", "databricks-sdk"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/terrametria/_version.py"

[tool.hatch.envs.default]
python = "3.10"
# we only needs these dependencies for development
dependencies = [
  "pyspark",
  "delta-spark",
  "pytest",
  "pandas",
  "numpy",
  "requests",
  "httpx",
]

[project.scripts]
loader = "terrametria.entrypoint:loader"

[tool.hatch.envs.default.scripts]
dev-frontend = "yarn --cwd src/frontend dev"
dev-backend = "uvicorn terrametria.backend:app --reload --host=0.0.0.0 --port=6006"
deploy-and-run = [
  'databricks bundle deploy --var="catalog=ivt" --var="schema=terrametria"',
  "databricks bundle run terrametria"
]